+++ webClient/src/app/editor/code-editor/monaco/monaco.component.ts
@@ -111,6 +111,10 @@ export class MonacoComponent implements OnInit, OnChanges {
       this.showEditor = !this.monacoService.spawnDiffViewer();
       this.showDiffViewer = !this.showEditor;
     });
+    
+    this.editorControl.refreshLayout.subscribe(() =>{
+      setTimeout(() => this.editor.layout(), 1);
+    });
   }
 
   focus(e: any) {
@@ -128,6 +132,10 @@ export class MonacoComponent implements OnInit, OnChanges {
           changes[input].currentValue['context'],
           changes[input].currentValue['reload'],
           changes[input].currentValue['line']);
+        //TODO: This is a workaround to instruct the editor to remeasure its container when switching from diff-viewer to code-editor
+        if(this.showDiffViewer) {
+          setTimeout(() => this.editor.layout(), 1);
+        }
         this.showEditor = true;
         this.showDiffViewer = false;
       }